% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_utils.R
\name{dot2pair}
\alias{dot2pair}
\title{Title dot bracket notation to a list of pairs}
\usage{
dot2pair(ss)
}
\arguments{
\item{ss}{sequence}
}
\value{
a list of pairs
}
\description{
Title dot bracket notation to a list of pairs
}
\examples{
rnamisc::dot2pair("((..))")
}
