% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_utils.R
\name{read_dssr}
\alias{read_dssr}
\title{Read SS pairs in json, generated by x3dna-dssar}
\usage{
read_dssr(x, paironly = FALSE)
}
\arguments{
\item{x}{either a json string or path to the json formatted SS file}

\item{paironly}{default FALSE, if TRUE, return pair positions}
}
\value{
a dataframe of SS pairs
}
\description{
Read SS pairs in json, generated by x3dna-dssar
}
